Use user and host value from webfinger when adding users
Before, when attempting to add a new user from a Mastodon instance with difference LOCAL_DOMAIN and WEB_DOMAIN, the addition would fail. The lookup of redirects in the host-meta and webfinger results would occur, but wouldn't be used throughout the process of adding a new user to the indexer.

Now, the determined user and host values are used throughout the process.